+++ src/Vector.hpp
@@ -7,27 +7,27 @@
 
 #include "Utilities.hpp"
 
-const float ArrowCoef = float(0.3);
-const float PerpendCoef = float(0.3);
+const double ArrowCoef = double(0.3);
+const double PerpendCoef = double(0.3);
 
 class Vector {
     public:
-        float x_;
-        float y_;
+        double x_;
+        double y_;
 
-        Vector(float x = 0, float y = 0) : x_(x), y_(y) {};
+        Vector(double x = 0, double y = 0) : x_(x), y_(y) {};
  
-        void Rotate(const float radian);
+        void Rotate(const double radian);
 
-        float LengthSquared() const;
-        float Length() const;
-        void Resize(const float newLength);
+        double LengthSquared() const;
+        double Length() const;
+        void Resize(const double newLength);
 
-        void operator *=(float scalar);
+        void operator *=(double scalar);
         void operator +=(const Vector& vectorToAdd);
         void operator -=(const Vector& vectorToSub);
 
-        Vector operator*(float scalar) const;
+        Vector operator*(double scalar) const;
         Vector operator+(const Vector& vectorToAdd) const;
         Vector operator-(const Vector& vectorToSub) const;
 
@@ -39,6 +39,6 @@ class Vector {
 std::ostream& operator<<(std::ostream& outStream, const Vector& curVector);
 
 Vector CreatePerpendicularVector(const Vector& vector);
-float GetDistanceSquared(const Vector& startPoint, const Vector& endPoint, const Vector& curPoint);
+double GetDistanceSquared(const Vector& startPoint, const Vector& endPoint, const Vector& curPoint);
 
-float OrientSquare(const Vector& a, const Vector& b, const Vector& c);
+double OrientSquare(const Vector& a, const Vector& b, const Vector& c);
